Directory updates for deployment
