+++ .github/workflows/app-service-dev.yml
@@ -1,16 +1,17 @@
 name: "DEV - Backend App Service"
 
 on:
-    workflow_dispatch:
-    push:
-        branches:
-            - main
-        paths:
-            - api/**
-    
+  workflow_dispatch:
+  push:
+    branches:
+      - main
+    paths:
+      - dex-portal-api/**
+
 jobs:
-    remote-trigger:
-        uses: ./.github/workflows/remote-trigger-template.yml
-        with:
-            WORKFLOW: app-service-dev.yml
-        secrets: inherit
+  remote-trigger:
+    uses: ./.github/workflows/remote-trigger-template.yml
+    with:
+      WORKFLOW: app-service-dev.yml
+    secrets: inherit
+
