+++ api/.gitignore
@@ -3,13 +3,6 @@ build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
-src/main/resources/application.properties
-.env
-.env.sh
-vars.sh
-api/src/main/resources/static
-
-!.env.example
 
 ### STS ###
 .apt_generated
@@ -41,3 +34,8 @@ out/
 
 ### VS Code ###
 .vscode/
+**/.DS_Store
+
+.env
+kls_database.db
+docs/
