+++ docker-compose.yml
@@ -3,10 +3,10 @@ version: "2.4"
 services:
   dexportalapi:
     build:
-      context: ./dex-portal-api
+      context: ./api
       dockerfile: Dockerfile
     container_name: dex-portal-api
     ports:
       - 8080:8080
     env_file:
-      - ./dex-portal-api/.env
+      - ./api/.env
